Refactor StudentProfile component and extract StudentProfileContent for better modularity
- Removed unused imports and components from studentProfile.jsx
- Introduced loading and error states for improved user experience
- Implemented a new StudentProfileContent component to encapsulate student data display logic
- Moved data processing functions to a new studentDataProcessor.js utility file
- Updated student data fetching and processing logic to handle admin selection
- Enhanced UI with loading indicators and error alerts
